Fix Rect fields type from int to double

+++ src/Blorc.Core/Services/Interop/DocumentFunctionsInterop.cs
@@ -29,16 +29,16 @@
 
     public class Rect
     {
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
-        public int Height { get; set; }
+        public double Height { get; set; }
 
-        public int Width { get; set; }
+        public double Width { get; set; }
 
-        public int Bottom { get; set; }
+        public double Bottom { get; set; }
 
-        public int Top { get; set; }
+        public double Top { get; set; }
     }
 }
